Implement proof of concept worker, job, and recorder
Here we add a Job that accepts a URL and JSON arguments, then calls
that URL with those arguments. For the moment, we just print the URL
to the console.

A Route for accepting a task and writing it to the database.

And a worker that runs a single queue with 100 workers, processing
work items that come in.